+++ private/model/api/operation.go
@@ -31,9 +31,10 @@ type Operation struct {
 
 	EventStreamAPI *EventStreamAPI
 
-	IsEndpointDiscoveryOp bool               `json:"endpointoperation"`
-	EndpointDiscovery     *EndpointDiscovery `json:"endpointdiscovery"`
-	Endpoint              *EndpointTrait     `json:"endpoint"`
+	IsEndpointDiscoveryOp  bool               `json:"endpointoperation"`
+	EndpointDiscovery      *EndpointDiscovery `json:"endpointdiscovery"`
+	Endpoint               *EndpointTrait     `json:"endpoint"`
+	IsHttpChecksumRequired bool               `json:"httpChecksumRequired"`
 }
 
 // EndpointTrait provides the structure of the modeled endpoint trait, and its
@@ -142,6 +143,11 @@ func (o *Operation) GetSigner() string {
 	return buf.String()
 }
 
+// HasAccountIDMemberWithARN returns true if an account id member exists for an input shape that may take in an ARN.
+func (o *Operation) HasAccountIDMemberWithARN() bool {
+	return o.InputRef.Shape.HasAccountIdMemberWithARN
+}
+
 // operationTmpl defines a template for rendering an API Operation
 var operationTmpl = template.Must(template.New("operation").Funcs(template.FuncMap{
 	"EnableStopOnSameToken": enableStopOnSameToken,
@@ -207,6 +213,12 @@ func (c *{{ .API.StructName }}) {{ .ExportedName }}Request(` +
 		{{ .GetSigner }}
 	{{- end }}
 
+	{{- if .HasAccountIDMemberWithARN }}
+		// update account id or check if provided input for account id member matches 
+		// the account id present in ARN
+		req.Handlers.Validate.PushFrontNamed(updateAccountIDWithARNHandler)
+	{{- end }}
+
 	{{- if .ShouldDiscardResponse -}}
 		{{- $_ := .API.AddSDKImport "private/protocol" }}
 		{{- $_ := .API.AddSDKImport "private/protocol" .API.ProtocolPackage }}
@@ -221,7 +233,7 @@ func (c *{{ .API.StructName }}) {{ .ExportedName }}Request(` +
 				)
 			{{- end }}
 
-			es := new{{ $esapi.Name }}()
+			es := New{{ $esapi.Name }}()
 			{{- if $esapi.Legacy }}
 				req.Handlers.Unmarshal.PushBack(es.setStreamCloser)
 			{{- end }}
@@ -267,44 +279,56 @@ func (c *{{ .API.StructName }}) {{ .ExportedName }}Request(` +
 	{{- end }}
 
 	{{- if .EndpointDiscovery }}
-		{{- if not .EndpointDiscovery.Required }}
-			if aws.BoolValue(req.Config.EnableEndpointDiscovery) {
-		{{- end }}
-		de := discoverer{{ .API.EndpointDiscoveryOp.Name }}{
-			Required: {{ .EndpointDiscovery.Required }},
-			EndpointCache: c.endpointCache,
-			Params: map[string]*string{
-				"op": aws.String(req.Operation.Name),
-				{{- range $key, $ref := .InputRef.Shape.MemberRefs -}}
-					{{- if $ref.EndpointDiscoveryID -}}
-						{{- if ne (len $ref.LocationName) 0 -}}
-							"{{ $ref.LocationName }}": input.{{ $key }},
-						{{- else }}
-							"{{ $key }}": input.{{ $key }},
+		// if custom endpoint for the request is set to a non empty string,
+		// we skip the endpoint discovery workflow.
+		if req.Config.Endpoint == nil || *req.Config.Endpoint == "" {
+			{{- if not .EndpointDiscovery.Required }}
+				if aws.BoolValue(req.Config.EnableEndpointDiscovery) {
+			{{- end }}
+			de := discoverer{{ .API.EndpointDiscoveryOp.Name }}{
+				Required: {{ .EndpointDiscovery.Required }},
+				EndpointCache: c.endpointCache,
+				Params: map[string]*string{
+					"op": aws.String(req.Operation.Name),
+					{{- range $key, $ref := .InputRef.Shape.MemberRefs -}}
+						{{- if $ref.EndpointDiscoveryID -}}
+							{{- if ne (len $ref.LocationName) 0 -}}
+								"{{ $ref.LocationName }}": input.{{ $key }},
+							{{- else }}
+								"{{ $key }}": input.{{ $key }},
+							{{- end }}
 						{{- end }}
 					{{- end }}
-				{{- end }}
-			},
-			Client: c,
-		}
-
-		for k, v := range de.Params {
-			if v == nil {
-				delete(de.Params, k)
+				},
+				Client: c,
 			}
-		}
 
-		req.Handlers.Build.PushFrontNamed(request.NamedHandler{
-			Name: "crr.endpointdiscovery",
-			Fn: de.Handler,
-		})
-		{{- if not .EndpointDiscovery.Required }}
+			for k, v := range de.Params {
+				if v == nil {
+					delete(de.Params, k)
+				}
 			}
-		{{- end }}
+
+			req.Handlers.Build.PushFrontNamed(request.NamedHandler{
+				Name: "crr.endpointdiscovery",
+				Fn: de.Handler,
+			})
+			{{- if not .EndpointDiscovery.Required }}
+				}
+			{{- end }}
+		}
 	{{- end }}
 
 	{{- range $_, $handler := $.CustomBuildHandlers }}
 		req.Handlers.Build.PushBackNamed({{ $handler }})
+	{{- end }}
+
+	{{- if .IsHttpChecksumRequired }}
+		{{- $_ := .API.AddSDKImport "private/checksum" }}
+		req.Handlers.Build.PushBackNamed(request.NamedHandler{
+			Name: "contentMd5Handler",
+			Fn: checksum.AddBodyContentMD5Handler,
+		})
 	{{- end }}
 	return
 }